unix,win: add uv_if_indextoname, uv_if_indextoiid
if_indextoname(...) is a function defined in <net/if.h>

It is used to convert an IPv6 scope_id into an interface
identifier string such as %eth0 or %lo

uv_if_indextoiid is a cross-platform implementation that
returns an IPv6 interface identifier.  On Unix/Linux
it calls if_indextoname, on Windows it uses snprintf
to return the numeric interface identifier as a string.

Adds documentation and an ASSERT to the relevant test case.

PR-URL: https://github.com/libuv/libuv/pull/1445